+++ lib/lightning/jobs/job.ex
@@ -54,6 +54,8 @@ defmodule Lightning.Jobs.Job do
     belongs_to :workflow, Workflow
     has_one :project, through: [:workflow, :project]
 
+    field :delete, :boolean, virtual: true
+
     timestamps()
   end
 
@@ -75,29 +77,38 @@ defmodule Lightning.Jobs.Job do
         :workflow_id,
         :trigger_id
       ])
-      |> validate_required([
-        :name,
-        :body,
-        :enabled,
-        :adaptor
-      ])
 
     change
     |> cast_assoc(:trigger,
       with: {Trigger, :changeset, [change |> get_field(:workflow_id)]}
     )
+    |> validate()
+  end
+
+  # DEPRECATED: Jobs are now created via the workflow, this function is only
+  # used when creating a Job via a Trigger.
+  # Uncomment if this causes issues before fully removing other changeset/3
+  # functions.
+  # def changeset(job, attrs, workflow_id) do
+  #   attrs = Map.put(attrs, :workflow_id, workflow_id)
+  #
+  #   job
+  #   |> changeset(attrs)
+  #   |> validate_required(:workflow_id)
+  # end
+
+  def validate(changeset) do
+    changeset
+    |> validate_required([
+      :name,
+      :body,
+      :enabled,
+      :adaptor
+    ])
     |> assoc_constraint(:trigger)
     |> assoc_constraint(:workflow)
     |> validate_length(:name, max: 100)
     |> validate_format(:name, ~r/^[a-zA-Z0-9_\- ]*$/)
-  end
-
-  def changeset(job, attrs, workflow_id) do
-    attrs = Map.put(attrs, :workflow_id, workflow_id)
-
-    job
-    |> changeset(attrs)
-    |> validate_required(:workflow_id)
   end
 
   @doc """