+++ lib/lightning_web/controllers/changeset_json.ex
@@ -1,0 +1,15 @@
+defmodule LightningWeb.ChangesetJSON do
+  @moduledoc """
+  Renders changesets as JSON.
+  """
+
+  def error(%{changeset: changeset}) do
+    %{
+      errors:
+        Ecto.Changeset.traverse_errors(
+          changeset,
+          &LightningWeb.CoreComponents.translate_error/1
+        )
+    }
+  end
+end