+++ lib/lightning_web/router.ex
@@ -47,6 +47,8 @@ defmodule LightningWeb.Router do
       :require_authenticated_user
     ]
 
+    resources "/provision", API.ProvisioningController, only: [:create, :show]
+
     resources "/projects", API.ProjectController, only: [:index, :show] do
       resources "/jobs", API.JobController, only: [:index, :show]
       resources "/runs", API.RunController, only: [:index, :show]